FIX:refactored projects structure
