+++ src/App.css
@@ -36,7 +36,6 @@ body{
   height:90vh
 }
 
-
 /* sidebar designs */
 
 .sidebarAvatar{
@@ -105,7 +104,6 @@ body{
   color:#4b134f
 }
 
-
 /* navbar */
 
 .navbar {
@@ -139,69 +137,6 @@ body{
 }
 
 
-/* about component  */
-
-.aboutContainer{
-  width:90%;
-  margin:20px auto;
-  padding:20px;
-  background-color: #dbdddf;
-}
-
-.aboutHeading {
-  margin:0px 0px 30px 0px;
-}
-
-.aboutIntro {
-  padding:0px 20px;
-}
-
-.cardData {
-  background-color:#ffffff;
-  padding:20px 20px 5px 20px;
-  margin:10px;
-  border-radius: 10px;
-}
-
-
-.experienceDataList {
-  padding:0;
-  list-style: none;
-}
-
-.experienceDataList li {
-  padding:0;
-  margin:0;
-}
-
-.companyImage {
-   margin:20px 0px
-}
-
-
-/* resume component design */
-.resumeContainer{
-  width:100%;
-  margin:20px auto;
-  padding:20px;
-  /* background-color: #dbdddf; */
-}
-
-.resumeRow {
-  margin:20px 0px;
-}
-
-.progressRow{
-  margin-top:50px;
-}
-
-.progressHeader{
-  margin:30px 0px 0px 0px
-}
-
-.skillSection {
-  margin:60px 0px
-}
 
 .bar{
   width:90%;
@@ -230,61 +165,6 @@ body{
 
 /* projects component design */
 
-.projectContainer {
-  height: 70vh;
-  padding-top: 30px;
-  overflow-y: auto;
-}
-
-.projectCardWrapper {
-  margin: 10px 0 20px 0;
-  width: 100%;
-  height: 100%;
-  min-height: 160px;
-}
-
-.projectCardImage {
-  width: 100%;
-  height: 70%;
-  border-radius: 30px;
-  box-shadow: 0 0 5px rgb(187, 183, 183);
-  transition: transform 0.3s linear;
-}
-
-.projectTitle {
-  margin:5px 0px;
-}
-
-.githubIcon {
-  font-size: 20px;
-  color:#c94b4b;
-}
-
-.projectCardImage:hover {
-  transform: scale(1.05);
-}
-
-.active {
-  background-image: linear-gradient(90deg,#c94b4b,#4b134f);
-  margin:0 5px;
-  font-size: 1rem;
-  cursor: pointer;
-  color:#fff;
-}
-
-.projectNavbar {
-  display: flex;
-}
-
-.projectNavbar>div {
-  margin:0 5px;
-  font-size: 1rem;
-  cursor: pointer;
-  background-color: #f1f2f3;
-  padding:0px 5px;
-  border-radius: 5px;
-}
-
 @media screen and (max-width:1199px) {
   .contact {
     width: 90%;
@@ -297,6 +177,8 @@ body{
     margin: 30px 0px;
   }
 }
+
+
 
 @media screen and (max-width:357px) {
   .contact{
