+++ src/components/projects/project.css
@@ -1,0 +1,65 @@
+
+.projectContainer {
+    height: 70vh;
+    padding-top: 30px;
+    overflow-y: auto;
+  }
+  
+  .projectCardWrapper {
+    margin: 10px 0 20px 0;
+    width: 100%;
+    height: 100%;
+    min-height: 160px;
+  }
+  
+  .projectCardImage {
+    width: 100%;
+    height: 70%;
+    border-radius: 30px;
+    box-shadow: 0 0 5px rgb(187, 183, 183);
+    transition: transform 0.3s linear;
+  }
+  
+  .projectTitle {
+    margin:5px 0px;
+  }
+  
+  .githubIcon {
+    font-size: 20px;
+    color:#c94b4b;
+  }
+  
+  .projectCardImage:hover {
+    transform: scale(1.05);
+  }
+  
+  .active {
+    background-image: linear-gradient(90deg,#c94b4b,#4b134f);
+    margin:0 5px;
+    font-size: 1rem;
+    cursor: pointer;
+    color:#fff;
+  }
+  
+  .projectNavbar {
+    display: flex;
+  }
+  
+  .projectNavbar>div {
+    margin:0 5px;
+    font-size: 1rem;
+    cursor: pointer;
+    background-color: #f1f2f3;
+    padding:0px 5px;
+    border-radius: 5px;
+  }
+  
+
+  @media screen and (max-width:474px) {
+    .projectNavbar {
+        display: block;
+    }
+    .projectNavbar>div {
+        margin:10px 0;
+    }
+  }
